+++ database/migrations/2026_05_19_133419_remove_name_from_users_table.php
@@ -12,8 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-
-            $table->dropColumn('name');
+            if (Schema::hasColumn('users', 'name')) {
+                $table->dropColumn('name');
+            }
         });
     }
 
@@ -23,8 +24,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-
-            $table->string('name');
+            if (!Schema::hasColumn('users', 'name')) {
+                $table->string('name');
+            }
         });
     }
 };
